requirements.txt
pillow
ultralytics
opencv-python
requests
